[search]: use larger integer types

+++ src/search.c
@@ -7,7 +7,6 @@
 static char *search_string = NULL;
 static bool search_forward = true;
 
-
 /* pass NULL to highlight previous search */
 static inline void highlight(Ui *ui, const char *string)
 {
@@ -51,8 +50,8 @@ static void search_next_forward(Ui *ui, Fm *fm, bool inclusive)
 
 	Dir *dir = fm_current_dir(fm);
 	highlight(ui, NULL);
-	for (uint16_t i = inclusive ? 0 : 1; i < dir->length; i++) {
-		const uint16_t ind = (dir->ind + i) % dir->length;
+	for (uint32_t i = inclusive ? 0 : 1; i < dir->length; i++) {
+		const uint32_t ind = (dir->ind + i) % dir->length;
 		if (strcasestr(file_name(dir->files[ind]), search_string)) {
 			fm_cursor_move_to_ind(fm, ind);
 			return;
@@ -68,8 +67,8 @@ static void search_next_backwards(Ui *ui, Fm *fm, bool inclusive)
 
 	Dir *dir = fm_current_dir(fm);
 	highlight(ui, NULL);
-	for (uint16_t i = inclusive ? 0 : 1 ; i < dir->length; i++) {
-		const uint16_t ind = (dir->ind - i + dir->length) % dir->length;
+	for (uint32_t i = inclusive ? 0 : 1 ; i < dir->length; i++) {
+		const uint32_t ind = (dir->ind - i + dir->length) % dir->length;
 		if (strcasestr(file_name(dir->files[ind]), search_string)) {
 			fm_cursor_move_to_ind(fm, ind);
 			return;
